Save params as Data in Task to make the whole task codable
Making the property itself would require usage of generics, which means
whenever a new task is created, callers would need to tell what kind of
params it is gonna use. It also percolates in the rest of code with
generics.

To avoid that, we have simple getter and setter that encodes the params
at first step itself.